object_detection/tensorflow_object_api/requirements.txt
requests
Pillow
tensorflow-gpu<2.0.0
numpy
joblib
dataclasses
tqdm